scripts/requirements_p1_tests.txt
psutil==5.8.0
netifaces==0.10.9
pyyaml==5.4.1
requests==2.25.1
python-dateutil==2.8.1
pywin32==228; sys_platform == 'win32' 